+++ test/orm-metadata.spec.js
@@ -1,21 +1,22 @@
 import {Metadata, OrmMetadata} from '../src/orm-metadata';
 import {DefaultRepository} from '../src/default-repository';
 
-describe('OrmMetadata', function () {
-  describe('.forTarget()', function () {
-    it('Should always return a Metadata instance', function () {
+describe('OrmMetadata', function() {
+  describe('.forTarget()', function() {
+    it('Should always return a Metadata instance', function() {
       var meta = OrmMetadata.forTarget({});
 
       expect(meta instanceof Metadata).toBe(true);
       expect(meta.metadata).toEqual({
         repository  : DefaultRepository,
         resource    : null,
+        endpoint    : null,
         name        : null,
         associations: {}
       });
     });
 
-    it('Should return the same instance if already defined', function () {
+    it('Should return the same instance if already defined', function() {
       var target = {};
       var meta   = OrmMetadata.forTarget(target);
 
@@ -24,6 +25,7 @@ describe('OrmMetadata', function () {
         repository  : DefaultRepository,
         resource    : null,
         name        : null,
+        endpoint    : null,
         associations: {}
       });
 
@@ -34,9 +36,9 @@ describe('OrmMetadata', function () {
   });
 });
 
-describe('Metadata', function () {
-  describe('.addTo()', function () {
-    it('Should create an array if no value exists.', function () {
+describe('Metadata', function() {
+  describe('.addTo()', function() {
+    it('Should create an array if no value exists.', function() {
       var meta = new Metadata();
 
       meta.addTo('foo', 'bar');
@@ -44,7 +46,7 @@ describe('Metadata', function () {
       expect(meta.fetch('foo')).toEqual(['bar']);
     });
 
-    it('Should create an array with the current value in it, if value exists.', function () {
+    it('Should create an array with the current value in it, if value exists.', function() {
       var meta = new Metadata();
 
       meta.put('foo', 'bacon');
@@ -53,7 +55,7 @@ describe('Metadata', function () {
       expect(meta.fetch('foo')).toEqual(['bacon', 'bar']);
     });
 
-    it('Should append to an array.', function () {
+    it('Should append to an array.', function() {
       var meta = new Metadata();
 
       meta.put('foo', ['actual', 'array']);
@@ -62,7 +64,7 @@ describe('Metadata', function () {
       expect(meta.fetch('foo')).toEqual(['actual', 'array', 'bar']);
     });
 
-    it('Should return self.', function () {
+    it('Should return self.', function() {
       var meta = new Metadata();
 
       meta.addTo('foo', 'bar');
@@ -71,8 +73,8 @@ describe('Metadata', function () {
     });
   });
 
-  describe('.put()', function () {
-    it('Should put a value.', function () {
+  describe('.put()', function() {
+    it('Should put a value.', function() {
       var meta = new Metadata();
 
       meta.put('foo', 'bar');
@@ -80,7 +82,7 @@ describe('Metadata', function () {
       expect(meta.fetch('foo')).toEqual('bar');
     });
 
-    it('Should put a value nested.', function () {
+    it('Should put a value nested.', function() {
       var meta = new Metadata();
 
       meta.put('foo', {});
@@ -89,7 +91,7 @@ describe('Metadata', function () {
       expect(meta.fetch('foo', 'bar')).toEqual('bat');
     });
 
-    it('Should create an object whet you put a value nested that does not exist yet.', function () {
+    it('Should create an object whet you put a value nested that does not exist yet.', function() {
       var meta = new Metadata();
 
       meta.put('foo', 'bar', 'bat');
@@ -97,15 +99,15 @@ describe('Metadata', function () {
       expect(meta.fetch('foo', 'bar')).toEqual('bat');
     });
 
-    it('Should return self.', function () {
+    it('Should return self.', function() {
       var meta = new Metadata();
 
       expect(meta.put('foo', 'bar')).toEqual(meta);
     });
   });
 
-  describe('.has()', function () {
-    it('Should return true if metadata has value for key.', function () {
+  describe('.has()', function() {
+    it('Should return true if metadata has value for key.', function() {
       var meta = new Metadata();
 
       meta.put('foo', 'bar');
@@ -113,13 +115,13 @@ describe('Metadata', function () {
       expect(meta.has('foo')).toEqual(true);
     });
 
-    it('Should return false if metadata does not have value for key.', function () {
+    it('Should return false if metadata does not have value for key.', function() {
       var meta = new Metadata();
 
       expect(meta.has('foo')).toEqual(false);
     });
 
-    it('Should return true if metadata has nested value for key.', function () {
+    it('Should return true if metadata has nested value for key.', function() {
       var meta = new Metadata();
 
       meta.put('foo', 'bar', 'bat');
@@ -127,15 +129,15 @@ describe('Metadata', function () {
       expect(meta.has('foo', 'bar')).toEqual(true);
     });
 
-    it('Should return false if metadata does not have nested value for keys.', function () {
+    it('Should return false if metadata does not have nested value for keys.', function() {
       var meta = new Metadata();
 
       expect(meta.has('foo', 'bar')).toEqual(false);
     });
   });
 
-  describe('.fetch()', function () {
-    it('Should return the value when exists', function () {
+  describe('.fetch()', function() {
+    it('Should return the value when exists', function() {
       var meta = new Metadata();
 
       meta.put('foo', 'bar');
@@ -143,13 +145,13 @@ describe('Metadata', function () {
       expect(meta.fetch('foo')).toEqual('bar');
     });
 
-    it('Should return null when value not exists', function () {
+    it('Should return null when value not exists', function() {
       var meta = new Metadata();
 
       expect(meta.fetch('foo')).toEqual(null);
     });
 
-    it('Should return the nested value when exists', function () {
+    it('Should return the nested value when exists', function() {
       var meta = new Metadata();
 
       meta.put('foo', 'bar', 'bat');
@@ -157,7 +159,7 @@ describe('Metadata', function () {
       expect(meta.fetch('foo', 'bar')).toEqual('bat');
     });
 
-    it('Should return null when the nested value not exists', function () {
+    it('Should return null when the nested value not exists', function() {
       var meta = new Metadata();
 
       expect(meta.fetch('foo', 'bar')).toEqual(null);