+++ test/resources/entity/with-endpoint.js
@@ -1,0 +1,8 @@
+import {Entity} from '../../../src/entity';
+import {resource, endpoint} from '../../../src/index';
+
+@resource()
+@endpoint('sx/alternative')
+export class WithEndpoint extends Entity {
+
+}